fixs init camera

+++ sources/camera/set_fov.c
@@ -6,7 +6,7 @@
 /*   By: jibanez <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/02/17 15:08:43 by jibanez           #+#    #+#             */
-/*   Updated: 2015/02/17 15:37:33 by jibanez          ###   ########.fr       */
+/*   Updated: 2015/02/18 13:22:21 by jibanez          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ t_camera	set_fov(t_camera camera, float fov)
 	float	adjacent_side;
 
 	tan_a = tan(((fov / 2) * PI) / 180.0);
-	oposite_side = (float) camera.viewplane_width / 2.0;
+	oposite_side = camera.viewplane_width / 2.0;
 	adjacent_side = 1 / (tan_a / oposite_side);
 	camera.view_distance = adjacent_side;
 	return (camera);
